feat(performance) Optimize render performance
