+++ components/Header.tsx
@@ -16,6 +16,8 @@ import { useDispatch } from 'react-redux';
 import { authActions } from '../store/auth';
 import { logoutAPI } from '../lib/api/auth';
 import { userActions } from '../store/user';
+import HeaderAuths from './HeaderAuths';
+import HeaderUserProfile from './HeaderUserProfile';
 
 const Container = styled.div`
   position: sticky;
@@ -120,20 +122,7 @@ const Container = styled.div`
 `;
 
 const Header: React.FC = () => {
-  const dispatch = useDispatch();
-  const { openModal, ModalPortal, closeModal } = useModal();
-  const user = useSelector((state) => state.user);
-
-  const [isUserMenuOpened, setIsUserMenuOpened] = useState(false);
-
-  const logout = async () => {
-    try {
-      await logoutAPI();
-      dispatch(userActions.initUser());
-    } catch (e) {
-      console.log(e.message);
-    }
-  };
+  const isLogged = useSelector((state) => state.user.isLogged);
 
   return (
     <Container>
@@ -143,75 +132,8 @@ const Header: React.FC = () => {
           <AirbnbLogoTextIcon />
         </a>
       </Link>
-      {!user.isLogged && (
-        <div className='header-auth-buttons'>
-          <button
-            type='button'
-            className='header-button header-sign-up-button'
-            onClick={() => {
-              dispatch(authActions.setAuthMode('signup'));
-              openModal();
-            }}
-          >
-            회원가입
-          </button>
-          <button
-            type='button'
-            className='header-button header-login-button'
-            onClick={() => {
-              dispatch(authActions.setAuthMode('login'));
-              openModal();
-            }}
-          >
-            로그인
-          </button>
-        </div>
-      )}
-      {user.isLogged && (
-        <OutsideClickHandler
-          onOutsideClick={() => {
-            if (isUserMenuOpened) {
-              setIsUserMenuOpened(false);
-            }
-          }}
-        >
-          <button
-            className='header-user-profile'
-            type='button'
-            onClick={() => {
-              setIsUserMenuOpened(!isUserMenuOpened);
-            }}
-          >
-            <HamburgerIcon />
-            <img
-              src={user.profileImage}
-              className='header-user-profile-image'
-              alt=''
-            />
-          </button>
-          {isUserMenuOpened && (
-            <ul className='header-usermenu'>
-              <li>숙소 관리</li>
-              <Link href='/room/register/building'>
-                <a
-                  role='presentation'
-                  onClick={() => setIsUserMenuOpened(false)}
-                >
-                  <li>숙소 등록하기</li>
-                </a>
-              </Link>
-              <div className='header-usermenu-divider' />
-              <li role='presentation' onClick={logout}>
-                로그아웃
-              </li>
-            </ul>
-          )}
-        </OutsideClickHandler>
-      )}
-
-      <ModalPortal>
-        <AuthModal closeModal={closeModal} />
-      </ModalPortal>
+      {!isLogged && <HeaderAuths />}
+      {isLogged && <HeaderUserProfile />}
     </Container>
   );
 };
