+++ components/auth/SignUpModal.tsx
@@ -235,6 +235,10 @@ const SignUpModal: React.FC<IProps> = ({ closeModal }) => {
     dispatch(authActions.setAuthMode('login'));
   };
 
+  const disabledMonths = ['월'];
+  const disabledDays = ['일'];
+  const disabledYears = ['년'];
+
   return (
     <Container onSubmit={onSubmitSignUp}>
       <CloseXIcon className='modal-close-x-icon' onClick={closeModal} />
@@ -325,7 +329,7 @@ const SignUpModal: React.FC<IProps> = ({ closeModal }) => {
         <div className='sign-up-modal-birthday-month-selector'>
           <Selector
             options={monthList}
-            disabledOptions={['월']}
+            disabledOptions={disabledMonths}
             defaultValue='월'
             value={birthMonth}
             onChange={onChangeBirthMonth}
@@ -335,7 +339,7 @@ const SignUpModal: React.FC<IProps> = ({ closeModal }) => {
         <div className='sign-up-modal-birthday-day-selector'>
           <Selector
             options={dayList}
-            disabledOptions={['일']}
+            disabledOptions={disabledDays}
             defaultValue='일'
             value={birthDay}
             onChange={onChangeBirthDay}
@@ -345,7 +349,7 @@ const SignUpModal: React.FC<IProps> = ({ closeModal }) => {
         <div className='sign-up-modal-birthday-year-selector'>
           <Selector
             options={yearList}
-            disabledOptions={['년']}
+            disabledOptions={disabledYears}
             defaultValue='년'
             value={birthYear}
             onChange={onChangeBirthYear}
