+++ components/common/Button.tsx
@@ -22,4 +22,4 @@ const Button: React.FC<IProps> = ({ children, ...props }) => {
   return <Container {...props}>{children}</Container>;
 };
 
-export default Button;
+export default React.memo(Button);
