+++ components/common/Input.tsx
@@ -89,4 +89,4 @@ const Input: React.FC<IProps> = ({
   );
 };
 
-export default Input;
+export default React.memo(Input);
