+++ components/common/Selector.tsx
@@ -68,4 +68,4 @@ const Selector: React.FC<IProps> = ({
   );
 };
 
-export default Selector;
+export default React.memo(Selector);
